feat: customer - create with customerNo

+++ src/main/java/com/iws_manager/iws_manager_api/repositories/CustomerRepository.java
@@ -29,5 +29,5 @@ public interface CustomerRepository extends JpaRepository<Customer, Long> {
     Optional<Customer> findById(Long id);
 
     @Query("SELECT MAX(c.customerno) FROM Customer c")
-    Long findMaxCustomerNo();
+    Integer findMaxCustomerNo();
 }
